查看个人订单列表功能

+++ app/src/main/java/com/example/smartcity/models/UserInfoModel.java
@@ -29,7 +29,7 @@ public class UserInfoModel {
         @SerializedName("idCard")
         public String idCard;
         @SerializedName("balance")
-        public int balance;
+        public float balance;
         @SerializedName("score")
         public int score;
     }
